requirements.txt
torch==2.5.0
torchvision==0.20.0
Pillow==10.4.0
pandas==2.2.2
streamlit>=1.16.0
opencv-python==4.10.0.84